Add print functionality to individual trends page
A bit rudimentary but I have tried HTML2Canvas and three other libraries. The issues lies in the fact that these graphs are n ot rendered in the DOM but instead are pulled from Mongo Graphs. @MichaelEHammond when you record make sure you try doing control p and clicking save as pdf as your default option when printing. Therefore when the button is pressed it does the same on live video.