Add latestActivity and archiveDate to patient record
In preparation for implementation of patient archive functionality, two fields were added to a patient's record in the database: latestActivity and archiveDate.

latestActivity records the latest date (and time) of occurrence of one of the following activities: the date the patient profile was created, the date  the patient profile was most recently edited, the date a calculation was added for the patient, or the date a calculation was edited for the patient.

acrhiveDate is exactly 60 days from the latestActivity date.