Implement fixes to route files
Unsure of the reason why the previous implementation caused an issue where six database operations would cause the app to break by preventing records to be loaded and the only remedy would be to reload the page. It may have to do with the server incorrectly sending a response to the client.