Extract calculation inputs out of calculation components
Removed the calculation inputs defined in each calculation component and defined all inputs of all calculations in one place instead. The values and setState functions for each input is passed into each component instead. In this way, calculations that share the same inputs do not need to be inputted more than once.